requirements.txt
Django==1.8.8
psycopg2==2.6.1
pytz==2015.2
django-celery==3.1.16
Celery==3.1.18
kombu==3.0.26
amqp==1.4.6
gunicorn==0.17.4
django-balancer==0.4
django-storages==1.1.8
django_compressor==1.5
python-memcached==1.52
redis==2.10.3
django-redis-cache==1.3.0
django-cache-machine==0.9.1
newrelic==2.44.0.36

